feat!: Boundary order checking (#164)
Implements #161 as described in
https://github.com/CQCL/hugr/discussions/1656.

Once this is merged, we should be able to optimise the convex checker to
only test the edges returned by `PortOrdering::missing_pairs`. When the
original order is stronger than the other in a rewrite, convex checking
can be skipped altogether.

BREAKING CHANGE: `Subgraph::new_subgraph` now takes a `Boundary` as
argument.

---------

Co-authored-by: Luca Mondada <72734770+lmondada@users.noreply.github.com>