feat: add Subgraph::copy_in_parent (#182)
closes #171.

As suggested there, using Subgraph to define the, erm, subgraph to copy
seemed good.
* Since Subgraph borrows its parent graph, that reference has to be the
ref via which the parent is mutated, so this meant implementing LinkView
and LinkMut for &mut G
* And also dropping some redundant `: Clone` bounds, since `&mut`s do
not implement Clone

The API seems slightly strange in that the Subgraph defines what to
copy, but of course then does not include the copy of that just made,
but I think that makes sense.

Note that `Subgraph` never implements `LinkMut` so we avoid the complex
case of the parent itself being a Subgraph :)