feat!: Make PortGraph generic on node and port types (#230)
Closes #213

PortGraph and MultiPortGraph are now generic on `NodeIndex`, `PortIndex`
and `PortOffset`. This allows the user to fix the size of these fields,
using the unsigned integer of their choice.

To smooth the transition to fully generic indices, we currently only
implement most features in this crate for `NodeIndex` and `PortIndex`
fixed to `u32`. Generic `PortOffset` is supported in `PortView` and all
related traits.

BREAKING CHANGE: `PortGraph` and `MultiPortGraph` now take generic type
parameters. The default choices correspond to the old types. `PortView`
has a new associated type `PortOffsetBase`.