fix: Serialise/Deserialise for indices is backwards compatible (#234)
https://github.com/CQCL/portgraph/pull/230 broke serialisation of
`NodeIndex`, `PortIndex` and `PortOffset`. This PR restores the previous
serialisation.

Should I mark this PR as breaking, even though it does not change any
APIs? Sorry for the double breakage. The previous simpler serialisation
was better, and this will also break less user code in the long run
(minus the once that have already upgraded).

EDIT: We are yanking 0.15, restoring the old serialisation format is
thus non-breaking.